requirements.txt
Flask_Migrate==2.7.0
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
PyJWT==1.7.1
Flask-Mail==0.9.1